+++ UniSpace/Domain/Model/OwnerHouseSummary.swift
@@ -18,7 +18,7 @@ protocol OwnerHouseSummary {
     var size: Int { get set }
     var numberOfViews: Int { get set }
     var numberOfBookmarks: Int { get set }
-    var starRating: Int { get set }
+    var starRating: Double { get set }
     var arrangingTeamCount: Int { get set }
     var formingTeamCount: Int { get set }
     var houseStatus: HouseStatus { get set }
